+++ diive/__init__.py
@@ -28,6 +28,7 @@ from diive.core.times.resampling import resample_to_monthly_agg_matrix as resamp
 from diive.core.times.times import TimestampSanitizer as TimestampSanitizer
 from diive.pkgs.analysis import DailyCorrelation as DailyCorrelation
 from diive.pkgs.analysis import daily_correlation as daily_correlation
+from diive.pkgs.analysis import GrangerCausality as GrangerCausality
 from diive.pkgs.analysis.decoupling import StratifiedAnalysis as StratifiedAnalysis
 from diive.pkgs.analysis.decoupling import StratifiedAnalysis as stratified_analysis
 from diive.pkgs.analysis.gapfinder import GapFinder as GapFinder
@@ -164,6 +165,7 @@ __all__ = [
     # Packages: Analyses
     'DailyCorrelation',
     'daily_correlation',
+    'GrangerCausality',
     'StratifiedAnalysis',
     'stratified_analysis',
     'GapFinder',